+++ src/components/new-test/Timer.tsx
@@ -1,7 +1,7 @@
 import { Stack, TextField, Typography } from '@mui/material';
 
 export const Timer = () => (
-  <Stack >
+  <Stack>
     <Typography typography='inputTitle'>Timer</Typography>
     <Stack direction='row' spacing={1} alignItems='center'>
       <TextField value='05' />
