+++ src/types/CreateQuestionPropsType.ts
@@ -53,4 +53,3 @@ export enum TypeOptions {
   single = 'single',
   multi = 'multi',
 }
-
